requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# YouTube & Google APIs
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
youtube-transcript-api>=0.6.1
yt-dlp>=2023.10.13

# AI & ML
google-generativeai>=0.3.0

# Text-to-Speech
elevenlabs>=0.2.0

# Media Processing
Pillow>=10.0.0
moviepy>=1.0.3
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Arabic Text Processing
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# CLI & UI
typer>=0.9.0
rich>=13.5.0
questionary>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.7.0
pylint>=2.17.0
mypy>=1.5.0
isort>=5.12.0

# Type Stubs
types-requests>=2.31.0
types-python-dateutil>=2.8.0

